refactor(board): adjust outlets
