+++ src/pages/board/outlets/financialReporting/Requirements/index.tsx
