+++ src/routes/board.tsx
@@ -4,14 +4,12 @@ import { ErrorPage } from "@components/layout/ErrorPage";
 import { Board } from "@pages/board";
 import { BoardLayout } from "@pages/board/outlets/boardlayout";
 import { FinancialReporting } from "@pages/board/outlets/financialReporting";
-import { Requirements } from "@pages/board/outlets/boardlayout/Requirements";
-import { dataRequirements } from "@pages/board/outlets/boardlayout/Requirements/config";
-import { PromissoryNotes } from "@pages/board/outlets/boardlayout/PromissoryNotes";
-import { Postingvouchers } from "@pages/board/outlets/boardlayout/Postingvouchers";
+import { Requirements } from "@pages/board/outlets/financialReporting/Requirements";
+import { dataRequirements } from "@pages/board/outlets/financialReporting/Requirements/config";
+import { PromissoryNotes } from "@pages/board/outlets/financialReporting/PromissoryNotes";
+import { Postingvouchers } from "@pages/board/outlets/financialReporting/Postingvouchers";
 import { Approvals } from "@pages/board/outlets/financialReporting/Approvals";
-import { ToDo } from "@pages/board/outlets/boardlayout/ToDo";
-import { infoIcon } from "@pages/board/outlets/boardlayout/ToDo/config";
-import { CreditProfileInfo } from "@src/pages/board/outlets/creditProfileInfo";
+import { CreditProfileInfo } from "@pages/board/outlets/creditProfileInfo";
 
 function BoardRoutes() {
   return (
@@ -26,7 +24,6 @@ function BoardRoutes() {
               promissoryNotes={<PromissoryNotes />}
               postingVouchers={<Postingvouchers />}
               approvals={<Approvals />}
-              toDo={<ToDo icon={infoIcon} />}
             />
           }
         />
